修改跳轉頁面bug
