+++ app/src/main/java/com/example/idiotchefassistant/resultBlock/ResultRepository.kt
@@ -1,5 +1,15 @@
 package com.example.idiotchefassistant.resultBlock
+import android.util.Log
+import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import com.example.idiotchefassistant.itemBlock.IngredientItem
+import com.example.idiotchefassistant.recipeBlock.RecipeItem
+import okhttp3.MultipartBody
+import okhttp3.RequestBody.Companion.asRequestBody
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+import java.io.File
 import java.util.concurrent.Executors
 
 class ResultRepository {
@@ -19,6 +29,30 @@ class ResultRepository {
 
     fun getData(): ResultData? {
         return nowData.value
+    }
+
+    fun searchByIids(offset: Int, iids: List<Int>): LiveData<List<RecipeItem>> {
+        val liveData = MutableLiveData<List<RecipeItem>>()
+        resultSearchService.searchByIid(offset, iids).enqueue(object: Callback<List<RecipeItem>> {
+            override fun onResponse(
+                call: Call<List<RecipeItem>>,
+                response: Response<List<RecipeItem>>
+            ){
+                if(response.isSuccessful){
+                    liveData.postValue(response.body())
+                }
+                else{
+                    Log.i("searchByIds","Failed:${response.code()} ${response.message()}")
+                    liveData.postValue(emptyList())
+                }
+            }
+
+            override fun onFailure(call: Call<List<RecipeItem>>, t: Throwable) {
+                Log.e("searchByIds", "API call failed: ${t.message}")
+                liveData.postValue(emptyList())
+            }
+        })
+        return liveData
     }
 
     interface OnTaskFinish{
